Copy xtask version bumper from nrf-hal
Minor changes to accomodate the slightly different CHANGELOG format.
