+++ CHANGELOG.md
@@ -26,5 +26,5 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Fix rustdoc warnings
 - Upgrade nrf51-hal to 0.12.1
 
-[Unreleased]: https://github.com/therealprof/microbit/compare/v0.9.0...HEAD
-[0.9.0]: https://github.com/therealprof/microbit/compare/v0.8.0...v0.9.0
+[Unreleased]: https://github.com/nrf-rs/microbit/compare/v0.9.0...HEAD
+[0.9.0]: https://github.com/nrf-rs/microbit/compare/v0.8.0...v0.9.0
