+++ microbit-common/src/lib.rs
@@ -1,6 +1,7 @@
 //! microbit contains everything required to get started with the use of Rust
 //! to create firmwares for the fabulous [BBC micro:bit](https://microbit.org)
 //! microcontroller board.
+#![doc(html_root_url = "https://docs.rs/microbit-common/0.9.0")]
 #![no_std]
 #![deny(missing_docs)]
 #![allow(non_camel_case_types)]
