+++ microbit/Cargo.toml
@@ -28,3 +28,4 @@ license = "0BSD"
 [dependencies.microbit-common]
 path = "../microbit-common"
 features = ["v1"]
+version = "=0.9.0"
